Separated out the struct allocation/initialization from specializer to mem_initializer in S_TO_F operator
